fix: update folder sync interval handling in AlarmReceiver
- Changed the retrieval of the folder sync interval from SharedPreferences to use getLong() instead of getInt(), ensuring proper handling of the interval value as a Long before converting to Int.
- Updated the APK build to reflect these changes.

+++ mobile-app/android/app/src/main/kotlin/com/szurubooru/szuruqueue/FolderSyncAlarmReceiver.kt
@@ -80,7 +80,7 @@ class FolderSyncAlarmReceiver : BroadcastReceiver() {
         try {
             // Read sync interval from SharedPreferences
             val prefs = context.getSharedPreferences("FlutterSharedPreferences", Context.MODE_PRIVATE)
-            val intervalSeconds = prefs.getInt("flutter.folderSyncIntervalSeconds", 900)
+            val intervalSeconds = prefs.getLong("flutter.folderSyncIntervalSeconds", 900L).toInt()
 
             Log.d(TAG, "Rescheduling next alarm with interval: ${intervalSeconds}s")
